Fix feed items blinking on change. Fix tooltip wrong position after navigation form comments or edit profile.
