+++ app/src/main/java/com/alexbezhan/instagram/activities/postdetails/PostDetailsActivity.kt
@@ -1,51 +1,17 @@
 package com.alexbezhan.instagram.activities.postdetails
 
-import android.arch.lifecycle.LifecycleOwner
-import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.Observer
-import android.arch.lifecycle.Transformations
 import android.content.Intent
 import android.os.Bundle
 import android.support.v7.widget.LinearLayoutManager
 import com.alexbezhan.instagram.R
 import com.alexbezhan.instagram.activities.BaseActivity
-import com.alexbezhan.instagram.activities.BaseViewModel
-import com.alexbezhan.instagram.activities.asFeedPost
-import com.alexbezhan.instagram.activities.home.DefaultFeedPostListener
+import com.alexbezhan.instagram.activities.disableChangeAnimation
 import com.alexbezhan.instagram.activities.home.FeedAdapter
-import com.alexbezhan.instagram.activities.home.FeedPostStats
-import com.alexbezhan.instagram.activities.home.FeedPostListener
 import com.alexbezhan.instagram.activities.home.comments.CommentsActivity
 import com.alexbezhan.instagram.models.FeedPost
 import com.alexbezhan.instagram.models.User
-import com.alexbezhan.instagram.utils.firebase.FirebaseHelper.currentUid
-import com.alexbezhan.instagram.utils.firebase.FirebaseHelper.database
-import com.alexbezhan.instagram.utils.livedata.FirebaseLiveData
 import kotlinx.android.synthetic.main.activity_post_details.*
-
-class PostDetailsViewModel : BaseViewModel(), FeedPostListener {
-    private val feedPostListener = DefaultFeedPostListener(onFailureListener)
-
-    private lateinit var postId: String
-    lateinit var post: LiveData<FeedPost>
-
-    fun start(postId: String) {
-        this.postId = postId
-        post = Transformations.map(
-                FirebaseLiveData(database.child("feed-posts").child(currentUid()!!).child(postId)),
-                {
-                    it.asFeedPost()!!
-                })
-    }
-
-    override fun observePostStats(postId: String, owner: LifecycleOwner,
-                                  observer: Observer<FeedPostStats>) =
-            feedPostListener.observePostStats(postId, owner, observer)
-
-    override fun toggleLike(currentUser: User, post: FeedPost) =
-            feedPostListener.toggleLike(currentUser, post)
-
-}
 
 class PostDetailsActivity : BaseActivity(), FeedAdapter.Listener {
     private val TAG = "PostDetailsActivity"
@@ -60,6 +26,7 @@ class PostDetailsActivity : BaseActivity(), FeedAdapter.Listener {
             setContentView(R.layout.activity_post_details)
 
             mAdapter = FeedAdapter(this)
+            post_recycler.disableChangeAnimation()
             post_recycler.layoutManager = LinearLayoutManager(this)
             post_recycler.adapter = mAdapter
             back_image.setOnClickListener { finish() }
